+++ docs/versionhistory.rst
@@ -3,6 +3,13 @@ Version history
 
 This library adheres to `Semantic Versioning <http://semver.org/>`_.
 
+**1.2.0**
+
+- Renamed the ``isasyncgeneratorfunction`` function to ``isasyncgenfunction`` to match the new
+  function in the ``inspect`` module in Python 3.6 (the old name still works though)
+- Updated ``isasyncgenfunction`` to recognize native asynchronous generator functions in Python 3.6
+  and above
+
 **1.1.3** (2016-09-05)
 
 - Fixed error when throwing an exception into an asynchronous generator when using asyncio's debug
@@ -24,4 +31,4 @@ This library adheres to `Semantic Versioning <http://semver.org/>`_.
 
 **1.0.0** (2016-04-08)
 
-Initial release.
+- Initial release